defillama_api to gcs/requirements.txt
google-cloud-storage
requests
pandas